+++ src/mesido/esdl/asset_to_component_base.py
@@ -207,6 +207,10 @@ class _AssetToComponentBase:
         self._edr_pipes = json.load(
             open(os.path.join(Path(__file__).parent, "_edr_pipes.json"), "r")
         )
+        # The default gas pipe database is based on the ASA pipe catalogue schedule standard (std).
+        self._gas_pipes = json.load(
+            open(os.path.join(Path(__file__).parent, "_gas_pipe_database.json"), "r")
+        )
 
     def convert(self, asset: Asset) -> Tuple[Type[_Model], MODIFIERS]:
         """
@@ -323,6 +327,88 @@ class _AssetToComponentBase:
                     conductivies_insulation = [x.matter.thermalConductivity for x in components]
 
         return inner_diameter, insulation_thicknesses, conductivies_insulation
+
+    def _gas_pipe_get_diameter_and_roughness(self, asset: Asset) -> Tuple[float, float]:
+        """
+        There are multiple ways to specify pipe properties like inner-diameter and
+        pipe material and thickness.  The user specified nominal diameter (DN size)
+        takes precedence over potential user specified innerDiameter (while logging
+        warnings when either of these are specified in combination with the pipe DN).
+        Similarly, for the wall roughness, the order of preference is; based on the
+        DN size from the default database in this repo, the wall_roughness attribute
+        from the ESDL asset or the default value.
+        Parameters
+        ----------
+        asset : Asset pipe object with itss properties from ESDL
+
+        Returns
+        -------
+        pipe inner diameter [m], wall roughness [m]
+        """
+
+        full_name = f"{asset.asset_type} '{asset.name}'"
+        if asset.attributes["innerDiameter"] and asset.attributes["diameter"].value > 0:
+            logger.warning(
+                f"{full_name}' has both 'innerDiameter' and 'diameter' specified. "
+                f"Diameter of {asset.attributes['diameter'].name} will be used."
+            )
+
+        edr_dn_size = None
+        if asset.attributes["diameter"].value > 0:
+            edr_dn_size = str(asset.attributes["diameter"].name)
+        elif not asset.attributes["innerDiameter"]:
+            logger.error(
+                f"{full_name}' has no DN size or innerDiameter specified, default is set to DN200. "
+            )
+            edr_dn_size = "DN200"
+            # TODO: ideally no default diameter size is set, test cases have to be updated for this
+            #  to be removed.
+
+        # NaN means the default values will be used
+        wall_roughness = math.nan
+
+        if edr_dn_size:
+            # Get insulation and diameter properties from EDR asset with this size.
+            edr_asset = self._gas_pipes[edr_dn_size]
+            inner_diameter = float(edr_asset["inner_diameter"])
+            wall_roughness = float(edr_asset["roughness"])
+        else:
+            assert asset.attributes["innerDiameter"]
+            inner_diameter = asset.attributes["innerDiameter"]
+            if asset.attributes["roughness"] > 0.0:
+                wall_roughness = float(asset.attributes["roughness"])
+            else:
+                wall_roughness = 1.0e-4
+
+        return inner_diameter, wall_roughness
+
+    def _cable_get_resistance(self, asset: Asset) -> Tuple:
+        """
+        Determines the resistance in ohm/m defined by the inverse of the electrical conductivity
+        in the cable's material properties. If no material is defined a default resistance of
+        1e-6 ohm/m is used.
+
+        Parameters
+        ----------
+        asset : Asset cable object with it's properties from ESDL
+
+        Returns
+        -------
+        resistance
+        """
+        default_res = 1e-6
+        material = asset.attributes["material"]
+        el_conductivity = None
+        if material:
+            el_conductivity = material.electricalConductivity
+        if not el_conductivity:
+            logger.warning(
+                f"Cable {asset.name} does not have a material with conductivity assigned,"
+                f" using default resistance of {default_res} ohm/m"
+            )
+        res_ohm_per_m = 1 / el_conductivity if el_conductivity else default_res
+
+        return res_ohm_per_m
 
     def _is_disconnectable_pipe(self, asset: Asset) -> bool:
         """